refactor(formats): support sibling tags and type coercion in elemental & clean up tests
- Wrap Elemental HTML input in a virtual `<root>` tag to parse top-level sibling elements (like `<link>` outside `<body>`) without requiring LLM format changes.
- Add attribute type coercion (boolean, number, integer) and fallback JSON-like array/object parsing to handle LLM format variations.
- Make the `experiments` configuration in evals conditional on `version == "1.0"` in both solver and scorer to avoid loading v1.0 schemas for older formats.
- Merge the redundant `test_refactoring.py` assertions into `test_formats.py` and unify catalog provider test structures.
- Delete the empty `eval/tests/conftest.py` file.
