+++ agent_sdks/python/a2ui_agent/src/a2ui/inference_formats/experimental/elemental/compiler.py
@@ -290,7 +290,7 @@ class ElementalCompiler:
         self.expr_parser = ElementalExpressionParser()
 
         # Pre-compute container tags for forgiving parsing
-        self.container_tags = {"body", "template"}
+        self.container_tags = {"root", "body", "template"}
         for comp_name in self.helper.components:
             properties = self.helper.get_component_properties(comp_name)
             has_slotted_children = False
@@ -354,12 +354,40 @@ class ElementalCompiler:
     ) -> dict:
         """Compiles A2UI Elemental HTML into standard A2UI v1.0 wire JSON."""
         escaped_html = _escape_nested_script_tags(html_text)
+        # Wrap in a virtual <root> container to handle multiple top-level
+        # siblings gracefully (such as `<link rel="catalog">` and `<body>`). The
+        # HTML parser expects a single root element; wrapping them here ensures
+        # they are parsed inside a single DOM tree. This tag is a parser-only
+        # helper and is not expected to be generated by the model or seen in the
+        # output.
+        wrapped_html = f"<root>{escaped_html}</root>"
+
         builder = DomBuilder(self.container_tags)
-        builder.feed(escaped_html)
-        root = builder.root
+        builder.feed(wrapped_html)
+        virtual_root = builder.root
+
+        if not virtual_root or not virtual_root.children:
+            raise ValueError("A2UI Elemental document is empty.")
+
+        # Find the actual active root element inside the virtual root
+        root = None
+        for child in virtual_root.children:
+            if child.tag in [
+                "body",
+                f"{TAG_PREFIX}delete-surface",
+                f"{TAG_PREFIX}call-function",
+            ]:
+                root = child
+                break
 
         if not root:
-            raise ValueError("A2UI Elemental document is empty.")
+            # If no explicit body/op tag, default to body if ui-components are present,
+            # or raise an error.
+            raise ValueError(
+                "A2UI Elemental document must have a <body>,"
+                f" <{TAG_PREFIX}delete-surface>, or <{TAG_PREFIX}call-function> root"
+                " element."
+            )
 
         if root.tag == "body":
             # If there is a standalone operation inside body, treat it as the root
@@ -419,38 +447,45 @@ class ElementalCompiler:
                 envelope["functionCallId"] = func_call_id
             return envelope
 
-        if root.tag != "body":
-            raise ValueError(
-                "A2UI Elemental document must have a <body>,"
-                f" <{TAG_PREFIX}delete-surface>, or <{TAG_PREFIX}call-function> root"
-                " element."
-            )
-
         # 1. Surface ID from body
         surface_id = root.attrs.get("id", surface_id)
 
-        # 2. Extract catalog ID and data model from children
+        # 2. Extract catalog ID and data model from children of both virtual_root and body
         catalog_id_from_link = ""
         data_model = {}
         remaining_children = []
 
+        def extract_metadata(nodes):
+            nonlocal catalog_id_from_link, data_model
+            for child in nodes:
+                if child.tag == "link" and child.attrs.get("rel") == "catalog":
+                    catalog_id_from_link = child.attrs.get("href", "")
+                elif (
+                    child.tag == "script"
+                    and child.attrs.get("type") == "application/json"
+                ):
+                    if "slot" not in child.attrs:
+                        try:
+                            data_model = json.loads(child.text.strip())
+                        except json.JSONDecodeError as e:
+                            raise ValueError(
+                                f"Invalid JSON in dataModel script: {e}"
+                            ) from e
+
+        extract_metadata(virtual_root.children)
+        extract_metadata(root.children)
+
+        # Collect component children (children of body that are not link or root dataModel script)
         for child in root.children:
             if child.tag == "link" and child.attrs.get("rel") == "catalog":
-                catalog_id_from_link = child.attrs.get("href", "")
-            elif (
-                child.tag == "script" and child.attrs.get("type") == "application/json"
+                continue
+            if (
+                child.tag == "script"
+                and child.attrs.get("type") == "application/json"
+                and "slot" not in child.attrs
             ):
-                if "slot" not in child.attrs:
-                    try:
-                        data_model = json.loads(child.text.strip())
-                    except json.JSONDecodeError as e:
-                        raise ValueError(
-                            f"Invalid JSON in root dataModel script: {e}"
-                        ) from e
-                else:
-                    remaining_children.append(child)
-            else:
-                remaining_children.append(child)
+                continue
+            remaining_children.append(child)
 
         if not catalog_id:
             catalog_id = catalog_id_from_link or self.helper.catalog.get(
@@ -615,6 +650,36 @@ class ElementalCompiler:
                     parsed_val = ""
             else:
                 parsed_val = self.expr_parser.parse(attr_val)
+
+            # Retrieve property schema for type coercion/coaxing
+            prop_schema = self.helper.get_property_schema(comp_name, prop_name)
+
+            if isinstance(parsed_val, str) and prop_schema:
+                prop_type = prop_schema.get("type")
+                # 1. Coerce boolean values
+                if prop_type == "boolean":
+                    if parsed_val.lower() == "true":
+                        parsed_val = True
+                    elif parsed_val.lower() == "false":
+                        parsed_val = False
+                # 2. Coerce number/integer values
+                elif prop_type in ["number", "integer"]:
+                    try:
+                        if prop_type == "integer":
+                            parsed_val = int(parsed_val)
+                        else:
+                            parsed_val = float(parsed_val)
+                    except ValueError:
+                        pass
+                # 3. Parse string arrays/objects fallback as expression if they start with [ or {
+                elif prop_type in ["array", "object"] or (
+                    parsed_val.startswith("[") or parsed_val.startswith("{")
+                ):
+                    try:
+                        # Wrap in braces virtually to compile using expression parser
+                        parsed_val = self.expr_parser.parse(f"{{{parsed_val}}}")
+                    except Exception:
+                        pass
 
             # Post-process expression value (events, functions, etc.)
             parsed_val = self._compile_value(
