+++ agent_sdks/python/a2ui_agent/tests/test_formats.py
@@ -18,6 +18,11 @@ from a2ui.schema.constants import VERSION_0_9
 from a2ui.inference_formats.transport import TransportFormat, TransportParser
 from a2ui.adk.a2a.part_converter import A2uiPartConverter
 from google.genai import types as genai_types
+from a2ui.inference_formats.experimental.express import ExpressFormat, ExpressParser
+from a2ui.inference_formats.experimental.elemental import (
+    ElementalFormat,
+    ElementalParser,
+)
 
 
 @pytest.fixture
@@ -95,3 +100,125 @@ def test_strategy_based_converters(test_catalog, monkeypatch):
     )
     parts_json = json_converter.convert(part_json)
     assert len(parts_json) == 1
+
+
+def test_supports_streaming_property(test_catalog):
+    from a2ui.schema.catalog import CatalogConfig
+    from a2ui.schema.catalog_provider import A2uiCatalogProvider
+
+    class MemoryCatalogProvider(A2uiCatalogProvider):
+
+        def __init__(self, schema):
+            self.schema = schema
+
+        def load(self):
+            return self.schema
+
+    config = CatalogConfig(
+        name="test_catalog",
+        provider=MemoryCatalogProvider(test_catalog.catalog_schema),
+    )
+
+    # 1. TransportFormat parser supports streaming
+    transport_fmt = TransportFormat(version=VERSION_0_9, catalogs=[config])
+    assert transport_fmt.supports_streaming is True
+    assert transport_fmt.parser.supports_streaming is True
+
+    # 2. ExpressFormat parser does not support streaming
+    express_fmt = ExpressFormat(catalog=test_catalog)
+    assert express_fmt.supports_streaming is False
+    assert express_fmt.parser.supports_streaming is False
+
+    # 3. ElementalFormat parser does not support streaming
+    elemental_fmt = ElementalFormat(catalog=test_catalog)
+    assert elemental_fmt.supports_streaming is False
+    assert elemental_fmt.parser.supports_streaming is False
+
+
+def test_process_chunk_raises_not_implemented(test_catalog):
+    express_parser = ExpressParser(test_catalog)
+    with pytest.raises(NotImplementedError) as exc_info:
+        express_parser.process_chunk("chunk")
+    assert "Streaming is not supported by ExpressParser" in str(exc_info.value)
+
+    elemental_parser = ElementalParser(test_catalog)
+    with pytest.raises(NotImplementedError) as exc_info:
+        elemental_parser.process_chunk("chunk")
+    assert "Streaming is not supported by ElementalParser" in str(exc_info.value)
+
+
+def test_decompiler_delegation(test_catalog):
+    # Verify Transport Decompiler
+    transport_fmt = TransportFormat(version=VERSION_0_9, catalogs=[])
+    payload = {"createSurface": {"surfaceId": "main"}}
+    direct_decompile = transport_fmt.decompiler.decompile(payload)
+    assert "createSurface" in direct_decompile
+    assert "main" in direct_decompile
+
+    # Verify Express Decompiler
+    express_fmt = ExpressFormat(catalog=test_catalog)
+    expr_decompiler = express_fmt.decompiler
+    envelope = {
+        "version": "v1.0",
+        "createSurface": {
+            "surfaceId": "main",
+            "components": [{
+                "id": "root",
+                "component": "Text",
+                "text": "Hello World",
+            }],
+        },
+    }
+    decompiled_dsl = expr_decompiler.decompile(envelope)
+    assert 'root = Text("Hello World")' in decompiled_dsl
+
+    # Verify wrap_decompiled_blocks implementation
+    assert (
+        transport_fmt.decompiler.wrap_decompiled_blocks(["{}", "{}"])
+        == "<a2ui-json>\n{}\n{}\n</a2ui-json>"
+    )
+    assert (
+        expr_decompiler.wrap_decompiled_blocks(["a = 1", "b = 2"])
+        == "<a2ui>\na = 1\nb = 2\n</a2ui>"
+    )
+
+    # Verify base class fallback
+    from a2ui.decompiler import Decompiler
+
+    class DummyDecompiler(Decompiler):
+
+        def decompile(self, val):
+            return ""
+
+    assert (
+        DummyDecompiler().wrap_decompiled_blocks(["a = 1", "b = 2"]) == "a = 1\nb = 2"
+    )
+
+    # Verify abstract PromptGenerator generate pass
+    from a2ui.prompt.generator import PromptGenerator
+
+    class DummyPromptGenerator(PromptGenerator):
+
+        def generate(self, *args, **kwargs):
+            return super().generate(*args, **kwargs)
+
+    assert DummyPromptGenerator().generate("role") is None
+
+    # Verify invalid catalog_id check
+    bad_catalog = A2uiCatalog(
+        version="1.0",
+        name="bad",
+        experiments=None,
+        s2c_schema={},
+        common_types_schema={},
+        catalog_schema={"catalogId": 12345},
+    )
+    from a2ui.core.exceptions import A2uiCatalogError
+
+    with pytest.raises(A2uiCatalogError) as ctx:
+        _ = bad_catalog.catalog_id
+    assert "catalogId is not a string" in str(ctx.value)
+
+    # Verify empty pruned components and messages fallback
+    assert test_catalog._with_pruned_components([]) is test_catalog
+    assert test_catalog._with_pruned_messages([]) is test_catalog
