+++ eval/a2ui_eval/scorers.py
@@ -50,7 +50,11 @@ def a2ui_scorer(version: str):
         resolved_catalog_path = str(GIT_ROOT / catalog_path)
 
         catalog_config = CatalogConfig.from_path("basic_catalog", resolved_catalog_path)
-        transport_format = TransportFormat(version=version, catalogs=[catalog_config])
+        transport_format = TransportFormat(
+            version=version,
+            catalogs=[catalog_config],
+            experiments={"version_1_0"} if version == "1.0" else None,
+        )
         catalog = transport_format.get_selected_catalog()
         validator = catalog.validator
